Fixed 32 bit windows build

+++ src/RingArrays.jl
@@ -71,7 +71,7 @@ end
 
 checkbounds(ring::RingArray) = true # because of warnings
 
-function checkbounds{T, N}(ring::RingArray{T, N}, indexes::UnitRange{Int64}...)
+function checkbounds{T, N}(ring::RingArray{T, N}, indexes::UnitRange{Int}...)
     all_indexes = product(indexes...)
     for i in all_indexes
         i = expand_index(ring, i...)
